+++ src/components/App/App.jsx
@@ -4,20 +4,24 @@ import { Container } from "semantic-ui-react";
 import { withAuthenticator } from "aws-amplify-react";
 
 import NavBar from "../Nav/NavBar";
-import HomePage from "../../HomePage";
+import Home from "../Home/Home";
 import Accounts from "../Accounts/AccountsList";
-import ProtectedApp from "./ProtectedApp";
 
-class App extends React.Component {
-  render() {
-    return (
-      <>
-        <Route exact path='/' component={HomePage} />
-        <Route exact path='/(.+)' render={() => <ProtectedApp />} />
-      </>
-    );
-  }
+function App() {
+  return (
+    <>
+      <NavBar />
+      <Container className='main'>
+        <Route exact path='/home' component={Home} />
+        <Route exact path='/accounts' component={Accounts} />
+        {/* <Route
+                exact
+                path='/transactions/new'
+                component={TransactionAdd}
+              /> */}
+      </Container>
+    </>
+  );
 }
 
-//export default withAuthenticator(App, { includeGreetings: false });
-export default App;
+export default withAuthenticator(App, { includeGreetings: false });
