+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 import Amplify from "aws-amplify";
 import aws_exports from "./aws-exports";
-import App from "./components/App/App.jsx";
+import App from "./App.jsx";
 import "./index.css";
 
 Amplify.configure(aws_exports);
